+++ apps/web-elrond/src/components/pagination/components/actions/hooks.ts
@@ -3,8 +3,8 @@ export const useTablePaginationActions = (props: {
   // backIconButtonProps?: any;
   count: number;
   // nextIconButtonProps?: any;
-  onChangePage: (event: React.MouseEvent<HTMLButtonElement> | null, page: number) => void;
-  handleChangeRowsPerPage: (selectedRowsPerPage: number) => void;
+  onPageChange: (event: React.MouseEvent<HTMLButtonElement> | null, page: number) => void;
+  handleRowsPerPageChange: (selectedRowsPerPage: number) => void;
   page: number;
   rowsPerPage: number;
   pageNeighbors?: 1 | 2;
@@ -13,29 +13,29 @@ export const useTablePaginationActions = (props: {
     count,
     page,
     rowsPerPage,
-    onChangePage,
+    onPageChange,
     pageNeighbors = 1,
-    handleChangeRowsPerPage,
+    handleRowsPerPageChange,
   } = props;
 
   const handleFirstPage = () => {
-    onChangePage(null, 0);
+    onPageChange(null, 0);
   };
 
   const handleNextPage = () => {
     if (page + 1 <= Math.ceil(count / rowsPerPage) - 1) {
-      onChangePage(null, page + 1);
+      onPageChange(null, page + 1);
     }
   };
 
   const handlePreviousPage = () => {
     if (page - 1 >= 0) {
-      onChangePage(null, page - 1);
+      onPageChange(null, page - 1);
     }
   };
 
   const handleLastPage = () => {
-    onChangePage(null, Math.ceil(count / rowsPerPage) - 1);
+    onPageChange(null, Math.ceil(count / rowsPerPage) - 1);
   };
 
   const getAvailablePages = () => {
@@ -67,7 +67,7 @@ export const useTablePaginationActions = (props: {
   };
 
   const handleRowOptionChange = (event: React.ChangeEvent<{ value: unknown }>) => {
-    handleChangeRowsPerPage(Number(event.target.value));
+    handleRowsPerPageChange(Number(event.target.value));
   };
 
   return {