+++ apps/web-nomic/src/screens/validators/components/list/components/desktop/index.tsx
@@ -1,20 +1,21 @@
-import React, { ReactNode } from 'react';
-import classnames from 'classnames';
-import numeral from 'numeral';
-import useTranslation from 'next-translate/useTranslation';
-import AutoSizer from 'react-virtualized-auto-sizer';
-import { VariableSizeGrid as Grid } from 'react-window';
-import Typography from '@material-ui/core/Typography';
-import { useGrid } from '@/hooks';
-import SortArrows from '@/components/sort_arrows';
+import chainConfig from '@/chainConfig';
 import AvatarName from '@/components/avatar_name';
 import InfoPopover from '@/components/info_popover';
-import { getValidatorStatus } from '@/utils/get_validator_status';
+import SortArrows from '@/components/sort_arrows';
+import { useGrid } from '@/hooks';
 import { useStyles } from '@/screens/validators/components/list/components/desktop/styles';
 import { fetchColumns } from '@/screens/validators/components/list/components/desktop/utils';
-import type { ItemType } from '@/screens/validators/components/list/types';
 import VotingPower from '@/screens/validators/components/list/components/voting_power';
 import VotingPowerExplanation from '@/screens/validators/components/list/components/voting_power_explanation';
+import type { ItemType } from '@/screens/validators/components/list/types';
+import { getValidatorStatus } from '@/utils/get_validator_status';
+import Typography from '@material-ui/core/Typography';
+import classnames from 'classnames';
+import useTranslation from 'next-translate/useTranslation';
+import numeral from 'numeral';
+import React, { ReactNode } from 'react';
+import AutoSizer from 'react-virtualized-auto-sizer';
+import { VariableSizeGrid as Grid } from 'react-window';
 
 const Desktop: React.FC<{
   className?: string;
@@ -28,13 +29,15 @@ const Desktop: React.FC<{
   const columns = fetchColumns();
 
   const { gridRef, columnRef, onResize, getColumnWidth, getRowHeight } = useGrid(columns);
+  const { tokenUnits, votingPowerTokenUnit } = chainConfig();
 
   const formattedItems = props.items.map((x, i): { [key: string]: ReactNode } => {
     const status = getValidatorStatus(x.inActiveSet, x.jailed, x.tombstoned);
     const percentDisplay = x.inActiveSet
       ? `${numeral(x.votingPowerPercent).format('0.[00]')}%`
       : '0%';
-    const votingPower = numeral(x.votingPower).format('0,0');
+    const { exponent } = tokenUnits[votingPowerTokenUnit];
+    const content = numeral(x.votingPower / 10 ** exponent).format('0,0');
 
     return {
       idx: `#${i + 1}`,
@@ -50,7 +53,7 @@ const Desktop: React.FC<{
         <VotingPower
           percentDisplay={percentDisplay}
           percentage={x.votingPowerPercent}
-          content={votingPower}
+          content={content}
           topVotingPower={x.topVotingPower ?? false}
         />
       ),