+++ packages/ui/src/screens/proposal_details/components/overview/index.tsx
@@ -4,12 +4,14 @@ import Name from '@/components/name';
 import SingleProposal from '@/components/single_proposal';
 import { useProfileRecoil } from '@/recoil/profiles';
 import { readDate } from '@/recoil/settings';
+import CommunityPoolSpend from '@/screens/proposal_details/components/overview/components/community_pool_spend';
 import ParamsChange from '@/screens/proposal_details/components/overview/components/params_change';
 import SoftwareUpgrade from '@/screens/proposal_details/components/overview/components/software_upgrade';
 import { useStyles } from '@/screens/proposal_details/components/overview/styles';
 import type { OverviewType } from '@/screens/proposal_details/types';
 import { getProposalType } from '@/screens/proposal_details/utils';
 import dayjs, { formatDayJs } from '@/utils/dayjs';
+import { formatNumber, formatToken } from '@/utils/format_token';
 import Divider from '@material-ui/core/Divider';
 import Typography from '@material-ui/core/Typography';
 import classnames from 'classnames';
@@ -30,7 +32,18 @@ const Overview: React.FC<{ overview: OverviewType } & ComponentDefault> = ({
   const type = getProposalType(R.pathOr('', ['@type'], overview.content));
 
   const proposer = useProfileRecoil(overview.proposer);
+  const recipient = useProfileRecoil(overview?.content?.recipient);
   const proposerMoniker = proposer ? proposer?.name : overview.proposer;
+  const recipientMoniker = recipient ? recipient?.name : overview?.content?.recipient;
+  const amountRequested = overview.content?.amount
+    ? formatToken(overview.content?.amount[0]?.amount, overview.content?.amount[0]?.denom)
+    : null;
+  const parsedAmountRequested = amountRequested
+    ? `${formatNumber(
+        amountRequested.value,
+        amountRequested.exponent
+      )} ${amountRequested.displayDenom.toUpperCase()}`
+    : '';
 
   const getExtraDetails = () => {
     let extraDetails = null;
@@ -53,6 +66,19 @@ const Overview: React.FC<{ overview: OverviewType } & ComponentDefault> = ({
             height={R.pathOr('0', ['plan', 'height'], overview.content)}
             info={R.pathOr('', ['plan', 'info'], overview.content)}
             name={R.pathOr('', ['plan', 'name'], overview.content)}
+          />
+        </>
+      );
+    } else if (type === 'communityPoolSpendProposal') {
+      extraDetails = (
+        <>
+          <Typography variant="body1" className="label">
+            {t('content')}
+          </Typography>
+          <CommunityPoolSpend
+            recipient={overview?.content?.recipient}
+            recipientMoniker={recipientMoniker}
+            amountRequested={parsedAmountRequested}
           />
         </>
       );